Fix errors due to integer type expected
There are several places where an integer variable is expected but
a float is given.  In particular, NumPy arrays require that indices
be integers.  Floats may occur because of numerical operations like
round() or because data is loaded (e.g. loadmat()).

+++ opensauce/soundfile.py
@@ -71,7 +71,7 @@ class SoundFile(object):
 
     @property
     def ms_len(self):
-        ms_len = math.floor(len(self.wavdata) / self.fs * 1000)
+        ms_len = int(math.floor(len(self.wavdata) / self.fs * 1000))
         self.__dict__['ms_len'] = ms_len
         return ms_len
